Increment build number after current maximum build number
LMBD 0.9.22 was packaged for Linux and OSX but was not for Windows in
the past.

LMDB 0.9.22 has been recently packaged for Windows by the job triggered
by: 486095133fe93336d7c0d9184d923085dfe8054f.

Is it accessible on conda-forge, yet it is note resolvable for Windows,
one stil get: " nothing provides requested lmdb 0.9.22.*". For instance
see:
https://dev.azure.com/conda-forge/feedstock-builds/_build/results?buildId=715611&view=logs&j=171a126d-c574-5c8c-1269-ff3b989e923d&t=1183ba29-a0b5-5324-8463-2a49ace9e213&l=132

The last build number for the Linux and OSX builds for 0.9.22
surprisingly is 1000.

See: https://anaconda.org/conda-forge/lmdb/files

 	conda 	222.0 kB 	| osx-64/lmdb-0.9.22-h1de35cc_1000.tar.bz2 	 4 years and 6 months ago 	conda-forge 	615 	main gcc7 cf202003
	conda 	670.8 kB 	| linux-64/lmdb-0.9.22-h14c3975_1000.tar.bz2 	 4 years and 6 months ago 	conda-forge 	3841 	main gcc7 cf202003

This updates the build number to 1001 for uniformity, hopefully making
lmdb 0.9.22 resolvable for Windows.

Signed-off-by: Julien Jerphanion <git@jjerphan.xyz>